backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-channels==0.7.0
channels-redis==4.1.0
psycopg2-binary>=2.9
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
daphne==4.0.0
whitenoise==6.6.0
redis==5.0.1
drf-spectacular==0.27.2